requirements.txt
Django==4.2.2
django-bootstrap4==23.2
django-cors-headers==4.2.0
django-simpleui==2023.3.1
django-storages==1.13.2
djangorestframework==3.14.0
Django_redis==5.4.0
numpy==1.24.2
sympy==1.11.1

annotated-types==0.5.0
asgiref==3.7.2
autopep8==2.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cos-python-sdk-v5==1.9.25
crcmod==1.7
cushy-storage==1.2.2
cycler==0.11.0
distlib==0.3.7

drf-yasg==1.21.7
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
Faker==18.6.2
filelock==3.12.2
Flask==2.3.2
fonttools==4.39.3
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1

mysql-connector-python==8.0.32

numpy==1.24.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
platformdirs==3.10.0
protobuf==3.20.3
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycryptodome==3.18.0
pydantic==2.0
pydantic_core==2.0.1
PyGetWindow==0.0.9
Pygments==2.15.1
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.5
pytz==2023.3
PyYAML==6.0.1
reportlab==4.0.4
requests==2.30.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.24.3
Werkzeug==2.3.6
xmltodict==0.13.0
broadcast_service==1.3.1
stylecloud==0.5.2